news_project/RAG_test/.ipynb_checkpoints/requirements-checkpoint.txt
chromadb==0.5.5
sentence-transformers>=2.6.1
pypdf>=4.2.0
tiktoken>=0.7.0
rank-bm25>=0.2.2
openai>=1.44.0
python-dotenv>=1.0.1
